Confirmed that having the Nanaimo North account have its own OAuth flow doesn't require the user to re-authenticate every 2 weeks.
Also gave access to Chaldecott North, but I'm going to try something else first: making the calendar access app on the fields account, which is internal to VCB.

May 7 will be two weeks plus a day.  If I don't need to reauthenticate on that day, I can delete the calendar access projects from Nanaimo North and Chaldecott North.  If I do need to re-authenticate, I'll have to enable API access for the other 5 accounts and use separate clients for each.

Not a big deal since the authentication only needs to be done once.

So for the moment, we just have a single client from the fields account.