Changes for 13U A playoffs
13U A playoffs are round-robin, with all the seeds set as of May 14.  There are some if-necessary games on VCB fields, depending on how the round-robin portion of the schedule goes.  These games are scheduled for June 11 and 15.
The round-robin games were added to TeamSnap as league-controlled games instead of team-controlled, so I had to adjust the TeamSnap reader for that.  Playoff games have a label, "Playoffs", which is a good idea on Ada's part.  As 15U A and 18U AA seeds are set, we'll do the same thing with those divisions.
The if-necessary 13U A playoff games are added to the field calendar from Assignr since they aren't in TeamSnap.  I changed how the 13U A teams are mapped, I'm using the same dictionary as the other age groups now.  Some of the "Expos" games were added as "Vancouver Expos".  Instead of fixing this in Assignr, I've added an entry for "Vancouver Expos" to the dictionary, which should be safe.